+++ shared/shared/src/main/scala/io/github/edadma/petradb/Result.scala
@@ -21,6 +21,8 @@ case class PrepareResult(name: String)    extends Result
 case class DeallocateResult(name: String) extends Result
 case class CopyResult(rows: Int)              extends Result
 case class CreateSchemaResult(name: String)  extends Result
+case class CreateSequenceResult(name: String) extends Result
+case class DropSequenceResult(name: String)   extends Result
 case object BeginResult    extends Result
 case object CommitResult   extends Result
 case object RollbackResult extends Result